2-model-train/requirements.txt
mlflow==2.9.2
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
psycopg2-binary==2.9.9
